code cleanup (remove unused imports)
git-svn-id: http://astropew.googlecode.com/svn/trunk@150 d889380e-1d37-11de-8b5c-13accb87f508

+++ src/common/network/AckingFilter.java
@@ -53,7 +53,7 @@ public abstract class AckingFilter implements PacketFilter {
 		 * @return true if seq was inside the window
 		 */
 		public boolean received(byte seq) {
-			int diff = oldestPending - (int)seq;
+			int diff = oldestPending - seq;
 			if (diff < 0) {
 				diff += 256;
 			}
